[android] - move snapshot t NativeMapView
