+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/MapboxMap.java
@@ -1560,8 +1560,7 @@ public final class MapboxMap {
    */
   @UiThread
   public void snapshot(@NonNull SnapshotReadyCallback callback, @Nullable final Bitmap bitmap) {
-    // FIXME 12/02/2016
-    //mapView.snapshot(callback, bitmap);
+    nativeMapView.addSnapshotCallback(callback, bitmap);
   }
 
   /**
@@ -1571,8 +1570,7 @@ public final class MapboxMap {
    */
   @UiThread
   public void snapshot(@NonNull SnapshotReadyCallback callback) {
-    // FIXME 12/02/2016
-    //mapView.snapshot(callback, null);
+    snapshot(callback, null);
   }
 
   /**
